# platform_root/platform_api/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Exastro Platform API
  description: Exastro Platform API
  version: 2.9.0
servers:
- url: "https://{host}"
  description: API endpoint
  variables:
    host:
      default: localhost
- url: "http://{host}"
  description: API endpoint
  variables:
    host:
      default: localhost
- url: "{protocol}://{host}:{port}"
  description: API endpoint (Development)
  variables:
    protocol:
      default: https
      enum:
      - http
      - https
    host:
      default: localhost
    port:
      default: "30080"
# security:
# - BearerAuth: []
# - BasicAuth: []
...
paths:
...
  /api/{organization_id}/platform/workspaces/{workspace_id}/conversations:
    get:
      tags:
      - AIAssistantService
      summary: 会話一覧を取得
      description: ユーザーの会話一覧を取得する
      operationId: list_conversations
      parameters:
      - name: organization_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: status
        in: query
        description: ステータスフィルター
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - active
          - closed
          - archived
      - name: limit
        in: query
        description: 取得件数
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 50
      - name: offset
        in: query
        description: オフセット
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 0
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_25"
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/runtimeError"
      x-openapi-router-controller: controllers.ai_assistant_service_controller
    post:
      tags:
      - AIAssistantService
      summary: 会話を作成
      description: 新しいチャット会話を作成する
      operationId: create_conversation
      parameters:
      - name: organization_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/workspace_id_conversations_body"
        required: true
      responses:
        "200":
          description: 会話作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_26"
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/runtimeError"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/runtimeError"
      x-openapi-router-controller: controllers.ai_assistant_service_controller
  /api/{organization_id}/platform/workspaces/{workspace_id}/conversations/{conversation_id}/messages:
    get:
      tags:
      - AIAssistantService
      summary: メッセージ一覧を取得
      description: 会話のメッセージ履歴を取得する
      operationId: list_messages
      parameters:
      - name: organization_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: conversation_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: limit
        in: query
        description: 取得件数
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 100
      - name: offset
        in: query
        description: オフセット(message_seq基準)
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 0
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_27"
        "404":
          description: 会話が見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/runtimeError"
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/runtimeError"
      x-openapi-router-controller: controllers.ai_assistant_service_controller
    post:
      tags:
      - AIAssistantService
      summary: メッセージを送信
      description: ユーザーメッセージを送信し、AIから応答を取得する
      operationId: send_message
      parameters:
      - name: organization_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: conversation_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/conversation_id_messages_body"
        required: true
      responses:
        "200":
          description: メッセージ送信成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_28"
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/runtimeError"
        "404":
          description: 会話が見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/runtimeError"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/runtimeError"
      x-openapi-router-controller: controllers.ai_assistant_service_controller
...
components:
  schemas:
    responseOk:
      title: A successful base response.
      type: object
      properties:
        result:
          type: integer
          example: 200
        message:
          type: string
          example: SUCCESS
        ts:
          type: string
          format: date-time
      example:
        result: 200
        message: SUCCESS
        ts: 2000-01-23T04:56:07.000+00:00
...
    runtimeError:
      type: object
      properties:
        result:
          type: integer
          format: int32
          example: 500
        message:
          type: string
          example: Error Message
        ts:
          type: string
          format: date-time
        data:
          type: object
...
    conversationInfo:
      type: object
      properties:
        conversation_id:
          type: string
        title:
          type: string
        status:
          type: string
          enum:
          - active
          - closed
          - archived
        current_token_count:
          type: integer
        message_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    messageInfo:
      type: object
      properties:
        message_id:
          type: string
        conversation_id:
          type: string
        message_seq:
          type: integer
        role:
          type: string
          enum:
          - user
          - assistant
          - system
        message_text:
          type: string
        ai_service_id:
          type: string
        ai_model_id:
          type: string
        token_count:
          type: integer
        created_at:
          type: string
          format: date-time
...
    inline_response_200_25:
      allOf:
      - $ref: "#/components/schemas/responseOk"
      - type: object
        properties:
          data:
            $ref: "#/components/schemas/inline_response_200_25_data"
    workspace_id_conversations_body:
      required:
      - title
      type: object
      properties:
        title:
          type: string
          description: 会話タイトル
          example: Bedrockとのチャット
    inline_response_200_26:
      allOf:
      - $ref: "#/components/schemas/responseOk"
      - type: object
        properties:
          data:
            $ref: "#/components/schemas/inline_response_200_26_data"
    inline_response_200_27:
      allOf:
      - $ref: "#/components/schemas/responseOk"
      - type: object
        properties:
          data:
            $ref: "#/components/schemas/inline_response_200_27_data"
    conversation_id_messages_body:
      required:
      - message
      type: object
      properties:
        message:
          type: string
          description: ユーザーメッセージ
          example: こんにちは、Bedrockのテストです
        model_id:
          type: string
          description: 使用するモデルID(省略時はデフォルト)
          example: anthropic.claude-3-5-sonnet-20240620-v1:0
        ai_service_id:
          type: string
          description: AIサービスID(省略時は'bedrock')
          default: bedrock
    inline_response_200_28:
      allOf:
      - $ref: "#/components/schemas/responseOk"
      - type: object
        properties:
          data:
            $ref: "#/components/schemas/inline_response_200_28_data"
...
    inline_response_200_25_data:
      type: object
      properties:
        conversations:
          type: array
          items:
            $ref: "#/components/schemas/conversationInfo"
        count:
          type: integer
        total_count:
          type: integer
    inline_response_200_26_data:
      type: object
      properties:
        conversation_id:
          type: string
        title:
          type: string
        status:
          type: string
    inline_response_200_27_data:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/messageInfo"
        count:
          type: integer
        conversation_id:
          type: string
    inline_response_200_28_data_usage:
      type: object
      properties:
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        total_tokens:
          type: integer
    inline_response_200_28_data:
      type: object
      properties:
        conversation_id:
          type: string
        user_message_id:
          type: string
        assistant_message_id:
          type: string
        content:
          type: string
          description: AIの応答
        usage:
          $ref: "#/components/schemas/inline_response_200_28_data_usage"
...
  securitySchemes:
    BearerAuth:
      type: http
      description: Credentials or access token for API
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_BearerAuth
    BasicAuth:
      type: http
      description: Value converted from ID/Password in base64
      scheme: basic
      x-basicInfoFunc: swagger_server.controllers.authorization_controller.check_BasicAuth